ai-agent-projects/SINGLE-AGENT/stateful-AI-agent/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "stateful-ai-agent"
version = "0.1.0"
description = "A small memory-enabled AI agent baseline using SQLite-backed conversation state."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.13.6",
    "python-dotenv>=1.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
